+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     - Entering an existing subject code redirects to the subject's main page
   - Subject pages now prompt users to select the appropriate study-level event for which they want to enter a subject's sheet
   - Subject page now organizes subject sheets and events in a persistent left hand navigation bar
+  - The new subject page now contains a link to subject timeline and settings
 - **General Changes**
   - Streamlined login system by removing alternate logins
   - Removed approval process for new user registration